Relocated signup use case to proper folder
