+++ src/use_case/signup/SignupDataAccessInterface.java
@@ -1,6 +1,6 @@
 
 
-package use_case_signup;
+package use_case.signup;
 
 import entity.User;
