+++ src/use_case/signup/SignupInputBoundary.java
@@ -1,5 +1,5 @@
 
-package use_case_signup;
+package use_case.signup;
 
 public interface SignupInputBoundary {
     void execute(SignupInputData signupInputData);
