+++ src/use_case/signup/SignupInputData.java
@@ -1,5 +1,5 @@
 
-package use_case_signup;
+package use_case.signup;
 
 public class SignupInputData {
     private final String username;
