+++ src/use_case/signup/SignupOutputData.java
@@ -1,5 +1,5 @@
 
-package use_case_signup;
+package use_case.signup;
 
 public class SignupOutputData {
     private final String username;
